added back option to the list, add, and delete menus
The sample code works by using the last_action function in Menu.py to go back to the last menu the user was, but we can't call this if the last action we did was not accessing a menu, like "back" because the function gets confused when we call a string when it expects a menu. To fix this, I separated the menus in menu_definitions.py.

Also, added more information to boilerplate so we can test easier